requirements.txt
asgiref==3.5.2
Django==4.1
sqlparse==0.4.2
requests==2.25.1
python-decouple==3.4
django-crispy-forms==1.14.0
